feat: Calculate total duration of lessons in Cours class
This commit adds a new method `getLessonsDuration()` to the `Cours` class. This method calculates the total duration of all the lessons associated with a course by iterating over the lessons and summing up their durations. The total duration is then returned. This method can be used to display the total duration of a course in the course grid template.