feat: Add average rating calculation to Cours class
This commit adds a new method `getAverageRating()` to the `Cours` class. This method calculates the average rating of a course by iterating over the associated ratings and calculating the total rating and number of ratings. If there are no ratings, it returns null. This method can be used to display the average rating of a course in the course grid template.